% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{dfm_predict_model}
\alias{dfm_predict_model}
\title{Predictive DFM}
\usage{
dfm_predict_model(
  y,
  x,
  n_factors,
  p_y,
  p_f,
  include_const,
  max_iter,
  tol,
  forecast_h
)
}
\description{
Uses factors from covariates X to predict target variable y. This is similar to FA-VAR but uses DFM factor
  extraction
}
