% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecod025ps1-package.R
\docType{package}
\name{ecod025ps1-package}
\alias{ecod025ps1}
\alias{ecod025ps1-package}
\title{ecod025ps1: ECOD025 - Problem Set 1 about VAR, FA-VAR and DFM models}
\description{
Using armadillo4r for a straightforward implementation of an EM algorithm and the Kalman filter for DFM.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://WEBSITE.COM}
  \item Report bugs at \url{https://github.com/USERNAME/PKGNAME/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
