% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{var_model}
\alias{var_model}
\title{Main VAR estimation function that returns everything}
\usage{
var_model(y, p, include_const, forecast_h)
}
\arguments{
\item{y}{Time series vector (T x 1)}

\item{p}{Lag order}

\item{include_const}{Include constant term in VAR}

\item{forecast_h}{Forecast horizon (0 for no forecast)}
}
\description{
Estimates a VAR model using OLS.
}
