% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{var_model}
\alias{var_model}
\title{Main VAR estimation function that returns everything}
\usage{
var_model(y, p, include_const, forecast_h)
}
\description{
Main VAR estimation function that returns everything
}
